+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+.. _v41-0-5:
+
+41.0.5 - 2023-10-24
+~~~~~~~~~~~~~~~~~~~
+
+* Updated Windows, macOS, and Linux wheels to be compiled with OpenSSL 3.1.4.
+* Added a function to support an upcoming ``pyOpenSSL`` release.
+
 .. _v41-0-4:
 
 41.0.4 - 2023-09-19